+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/app/presenter/SurveillanceCsvPresenter.java
@@ -121,6 +121,8 @@ public class SurveillanceCsvPresenter {
 		firstRow.add(surv.getType().getName());
 		if(surv.getRandomizedSitesUsed() != null) {
 			firstRow.add(surv.getRandomizedSitesUsed().toString());
+		} else {
+			firstRow.add("");
 		}
 		result.add(firstRow);