+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/domain/CertificationStatusEvent.java
@@ -1,0 +1,57 @@
+package gov.healthit.chpl.domain;
+
+
+
+
+public class CertificationStatusEvent {
+
+	private Long id;
+	private String eventDate;
+	private Long certificationStatusId;
+	private String certificationStatusName;
+	private Long lastModifiedUser;
+	private String lastModifiedDate;
+	
+	public CertificationStatusEvent(){}
+	
+	public Long getId() {
+		return id;
+	}
+	public void setId(Long id) {
+		this.id = id;
+	}
+	public String getEventDate() {
+		return eventDate;
+	}
+	public void setEventDate(String eventDate) {
+		this.eventDate = eventDate;
+	}
+	public Long getLastModifiedUser() {
+		return lastModifiedUser;
+	}
+	public void setLastModifiedUser(Long lastModifiedUser) {
+		this.lastModifiedUser = lastModifiedUser;
+	}
+	public String getLastModifiedDate() {
+		return lastModifiedDate;
+	}
+	public void setLastModifiedDate(String lastModifiedDate) {
+		this.lastModifiedDate = lastModifiedDate;
+	}
+
+	public Long getCertificationStatusId() {
+		return certificationStatusId;
+	}
+
+	public void setCertificationStatusId(Long certificationStatusId) {
+		this.certificationStatusId = certificationStatusId;
+	}
+
+	public String getCertificationStatusName() {
+		return certificationStatusName;
+	}
+
+	public void setCertificationStatusName(String certificationStatusName) {
+		this.certificationStatusName = certificationStatusName;
+	}
+}