+++ RELEASE_NOTES.md
@@ -11,7 +11,10 @@ _Date TBD_
 * Add API call to get decertified certified products with pageCount set to the total number of decertified products
 * Add certificationDateStart and certificationDateEnd as advanced search parameters
 * Change corrective action plan search parameters to new surveillance search parameters - NOT BACKWARDS COMPATIBLE
-* Add /surveillance/download[?all=true] endpoints to allow download of CSV file with all surveillance and download of CSV file with surveillance that resulted in nonconformities
+* Send questionable activity email when product owner changes (not during a merge)
+* Add /surveillance/download[?type=all|basic|nonconformities] endpoint to allow download of CSV file with all surveillance and download of CSV file with surveillance that resulted in nonconformities
+* Add lookup to new tables for certification status event history
+* Add date of last certification status change to certified product details
 
 ### Bugs Fixed
 * Changed transition of care calculations for the EHR Certification ID to more closely match the rule.